drop cf its own category select
previously, if a select cf's "categories are based on" is empty, we will show
an extra <select> to let user filter values by category, which is kinda
misleading as the extra select is not a real cf, so it's not searchable.

if user wants to use the category select, there is an upgrade step to turn
categories into real cfs.

Fixes: #30426